add .grok-context to git
This file is mostly word salad but is used to help maintain
context for ML-driven completion queries using grokker.  I'm
erring on the side of adding it to git, since I just now moved
a large swath of text into it, much of which I plan to summarize
and move back to the README via grokker queries.